@@ -54,9 +54,9 @@
         "@types/through2": "latest",
         "@types/travis-fold": "latest",
         "@types/xml2js": "^0.4.0",
-        "@typescript-eslint/eslint-plugin": "2.3.2",
-        "@typescript-eslint/experimental-utils": "2.3.2",
-        "@typescript-eslint/parser": "2.3.2",
+        "@typescript-eslint/eslint-plugin": "2.18.0",
+        "@typescript-eslint/experimental-utils": "2.18.0",
+        "@typescript-eslint/parser": "2.18.0",
         "async": "latest",
         "azure-devops-node-api": "^8.0.0",
         "browser-resolve": "^1.11.2",
@@ -65,10 +65,10 @@
         "chalk": "latest",
         "convert-source-map": "latest",
         "del": "5.1.0",
-        "eslint": "6.5.1",
-        "eslint-formatter-autolinkable-stylish": "1.0.3",
-        "eslint-plugin-import": "2.18.2",
-        "eslint-plugin-jsdoc": "15.9.9",
+        "eslint": "6.8.0",
+        "eslint-formatter-autolinkable-stylish": "1.1.1",
+        "eslint-plugin-import": "2.20.0",
+        "eslint-plugin-jsdoc": "21.0.0",
         "eslint-plugin-no-null": "1.0.2",
         "fancy-log": "latest",
         "fs-extra": "^6.0.1",
