@@ -37,23 +37,8 @@ namespace ts {
 
         function check(node: Node) {
             if (isJsFile) {
-                switch (node.kind) {
-                    case SyntaxKind.FunctionExpression:
-                        const decl = getDeclarationOfExpando(node);
-                        if (decl) {
-                            const symbol = decl.symbol;
-                            if (symbol && (symbol.exports && symbol.exports.size || symbol.members && symbol.members.size)) {
-                                diags.push(createDiagnosticForNode(isVariableDeclaration(node.parent) ? node.parent.name : node, Diagnostics.This_constructor_function_may_be_converted_to_a_class_declaration));
-                                break;
-                            }
-                        }
-                    // falls through if no diagnostic was created
-                    case SyntaxKind.FunctionDeclaration:
-                        const symbol = node.symbol;
-                        if (symbol.members && (symbol.members.size > 0)) {
-                            diags.push(createDiagnosticForNode(isVariableDeclaration(node.parent) ? node.parent.name : node, Diagnostics.This_constructor_function_may_be_converted_to_a_class_declaration));
-                        }
-                        break;
+                if (canBeConvertedToClass(node)) {
+                    diags.push(createDiagnosticForNode(isVariableDeclaration(node.parent) ? node.parent.name : node, Diagnostics.This_constructor_function_may_be_converted_to_a_class_declaration));
                 }
             }
             else {
@@ -192,5 +177,23 @@ namespace ts {
 
     function getKeyFromNode(exp: FunctionLikeDeclaration) {
         return `${exp.pos.toString()}:${exp.end.toString()}`;
+    }
+
+    function canBeConvertedToClass(node: Node): boolean {
+        if (node.kind === SyntaxKind.FunctionExpression) {
+            if (isVariableDeclaration(node.parent) && node.symbol.members?.size) {
+                return true;
+            }
+
+            const decl = getDeclarationOfExpando(node);
+            const symbol = decl?.symbol;
+            return !!(symbol && (symbol.exports?.size || symbol.members?.size));
+        }
+
+        if (node.kind === SyntaxKind.FunctionDeclaration) {
+            return !!node.symbol.members?.size;
+        }
+
+        return false;
     }
 }
