@@ -30,7 +30,7 @@ namespace Harness {
 
             // eslint-disable-next-line @typescript-eslint/no-this-alias
             const cls = this;
-            describe(`${this.kind()} code samples`, function(this: Mocha.ISuiteCallbackContext) {
+            describe(`${this.kind()} code samples`, function (this: Mocha.ISuiteCallbackContext) {
                 this.timeout(600_000); // 10 minutes
                 for (const test of testList) {
                     cls.runTest(typeof test === "string" ? test : test.file);
@@ -41,7 +41,7 @@ namespace Harness {
             // eslint-disable-next-line @typescript-eslint/no-this-alias
             const cls = this;
             const timeout = 600_000; // 10 minutes
-            describe(directoryName, function(this: Mocha.ISuiteCallbackContext) {
+            describe(directoryName, function (this: Mocha.ISuiteCallbackContext) {
                 this.timeout(timeout);
                 const cp: typeof import("child_process") = require("child_process");
 
@@ -58,9 +58,11 @@ namespace Harness {
                         if (!fs.existsSync(submoduleDir)) {
                             exec("git", ["clone", config.cloneUrl, directoryName], { cwd });
                         }
-                        exec("git", ["reset", "HEAD", "--hard"], { cwd: submoduleDir });
-                        exec("git", ["clean", "-f"], { cwd: submoduleDir });
-                        exec("git", ["pull", "-f"], { cwd: submoduleDir });
+                        else {
+                            exec("git", ["reset", "HEAD", "--hard"], { cwd: submoduleDir });
+                            exec("git", ["clean", "-f"], { cwd: submoduleDir });
+                            exec("git", ["pull", "-f"], { cwd: submoduleDir });
+                        }
 
                         types = config.types;
 
@@ -123,7 +125,7 @@ ${stripAbsoluteImportPaths(result.stderr.toString().replace(/\r\n/g, "\n"))}`;
 
             // eslint-disable-next-line @typescript-eslint/no-this-alias
             const cls = this;
-            describe(`${this.kind()} code samples`, function(this: Mocha.ISuiteCallbackContext) {
+            describe(`${this.kind()} code samples`, function (this: Mocha.ISuiteCallbackContext) {
                 this.timeout(cls.timeout); // 20 minutes
                 before(() => {
                     cls.exec("docker", ["build", ".", "-t", "typescript/typescript"], { cwd: IO.getWorkspaceRoot() }); // cached because workspace is hashed to determine cacheability
