@@ -343,7 +343,7 @@ namespace ts.server {
 
                 session.send = Session.prototype.send;
                 assert(session.send);
-                expect(session.send(msg)).to.not.exist; // eslint-disable-line no-unused-expressions
+                expect(session.send(msg)).to.not.exist; // eslint-disable-line @typescript-eslint/no-unused-expressions
                 expect(lastWrittenToHost).to.equal(resultMsg);
             });
         });
