@@ -229,15 +229,15 @@ declare namespace n1 {
 }
 declare namespace n2 {
     let a: "Hello";
-    let b: any;
+    let b: "Hello";
     let c: "Hello";
-    let d: any;
-    let e: any;
+    let d: "Hello";
+    let e: "Hello";
 }
 declare namespace n3 {
     let a: "Hello" | "World";
-    let b: any;
-    let c: any;
-    let d: any;
+    let b: "Hello" | "World";
+    let c: "Hello" | "World";
+    let d: "Hello" | "World";
     let e: "Hello" | "World";
 }
